results: carry alnout row length as int64_t (--rowlen 0)
With --rowlen 0 the alnout block width is the whole alignment:
results_show_alnout computed int rowlen = (int)(qseqlen + dseqlen) and passed
it to align_show's int alignwidth parameter, which is then stored in the
already-64-bit Alignment::width that sizes the q_line/a_line/d_line buffers
and bounds the putop walk. A combined query+target length above INT_MAX
(~2 GB) makes rowlen negative: the line buffers are sized from the (negative,
small) width while putop keeps advancing position.line, which never reaches
the negative width to reset — so it writes past the buffers. A size-vs-walk
mismatch and heap overflow gated on a >2 GB alignment.

Make rowlen int64_t (dropping the narrowing casts; opt_rowlen is already
int64_t) and widen align_show's alignwidth parameter to int64_t so the value
that sizes the buffers is the same one that bounds the walk. Output is
unchanged for normal inputs; multi-block splitting and --rowlen 0 verified.

+++ src/showalign.cc
@@ -306,7 +306,7 @@ auto align_show(std::FILE * output_handle,
                 int64_t const cigarlen,
                 int const numwidth,
                 int const namewidth,
-                int const alignwidth,
+                int64_t const alignwidth,
                 int const strand) -> void
 {
